requirements.txt
alpaca-py==0.7.1
numpy==1.23.5
qpython==2.0.0